Add auth to all /user endpoints

+++ app/user.js
@@ -1,6 +1,10 @@
 "use strict";
 
 module.exports = class App extends require("@appFactory") {
+  _permission(params, auth) {
+    return params.id === auth.credentials.user._id.toString();
+  }
+
   async post({ payload }) {
     const user = await this.mongo("User").create(payload);
     return { response: { user }, code: 201 };
@@ -13,6 +17,7 @@ module.exports = class App extends require("@appFactory") {
   }
 
   async patch({ params, auth, payload }) {
+    if (!this._permission(params, auth)) throw this.Boom.forbidden();
     const user = await this.mongo("User").findByIdAndUpdate(
       params.id,
       payload,
@@ -23,6 +28,7 @@ module.exports = class App extends require("@appFactory") {
   }
 
   async delete({ params, auth }) {
+    if (!this._permission(params, auth)) throw this.Boom.forbidden();
     const user = await this.mongo("User").findByIdAndDelete(params.id, {
       useFindAndModify: true,
     });
